first stage of refactoring
