+++ incl/HttpMessage.hpp
@@ -4,15 +4,6 @@
 #include <string>
 #include <map>
 
-struct HttpMessage {
-    std::string method;
-    std::string path;
-    std::string query;
-    std::string fragment;
-    std::string fieldName;
-    std::string fieldValue;
-    std::map<std::string, std::string> headers;
-    std::string body;
-};
+
 
 #endif // HTTPMESSAGE_HPP
